Create README.md
Overview:
SQL is a critical skill in data analysis and is key to working with structured data. In this repository, I’ve included projects where I apply SQL to clean raw data, perform exploratory data analysis (EDA), and extract valuable insights. The goal is to provide a clear, transparent view of how I approach data from a foundational level, starting with data cleaning to in-depth analysis.